refactor: move computeOverflowFit to lib, delete unused hook and apply state latch
Relocated pure calculation computeOverflowFit to src/lib/overflowFit.ts to keep hooks pure.

Moved unit tests to src/lib/overflowFit.test.ts.

Deleted unused useOverflowFit custom hook and its test file to honor YAGNI.

Implemented isMeasured state-latch to prevent Layout Shift flash in PopularPositionChips.tsx on load.
