+++ src/app/mentor-pool/PopularPositionChips.tsx
@@ -4,7 +4,7 @@ import { ChevronLeft, ChevronRight } from 'lucide-react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useRef, useState, useTransition } from 'react';
 
-import { computeOverflowFit } from '@/hooks/useOverflowFit';
+import { computeOverflowFit } from '@/lib/overflowFit';
 
 import { POPULAR_POSITIONS } from './data';
 import { buildHref, setSearchPattern } from './searchParams';
@@ -23,6 +23,7 @@ export default function PopularPositionChips() {
   const buttonRefs = useRef<(HTMLButtonElement | null)[]>([]);
   const scrollRef = useRef<HTMLDivElement | null>(null);
   const [visibleCount, setVisibleCount] = useState(POPULAR_POSITIONS.length);
+  const [isMeasured, setIsMeasured] = useState(false);
   const [canScrollLeft, setCanScrollLeft] = useState(false);
   const [canScrollRight, setCanScrollRight] = useState(true);
 
@@ -36,6 +37,7 @@ export default function PopularPositionChips() {
       defaultVisibleCount: POPULAR_POSITIONS.length,
     });
     setVisibleCount(computedCount);
+    setIsMeasured(true);
   }, []);
 
   useEffect(() => {
@@ -63,7 +65,9 @@ export default function PopularPositionChips() {
   };
 
   return (
-    <div className="relative">
+    <div
+      className={`relative transition-opacity duration-150 ${isMeasured ? 'opacity-100' : 'opacity-0'}`}
+    >
       <div
         ref={scrollRef}
         className="flex gap-2 overflow-x-auto [-ms-overflow-style:none] [scrollbar-width:none] xl:mx-auto xl:w-[846px] xl:justify-center xl:overflow-x-visible [&::-webkit-scrollbar]:hidden"
